backend/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
python-dotenv
pymongo
